requirements.txt
cbor2==5.5.1
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
observable==0.3.2
pubnub==4.8.1
pycryptodome==3.19.0
pycryptodomex==3.19.0
python-dotenv==1.0.0
requests==2.31.0
ringcentral==0.8.2
six==1.16.0
urllib3==2.0.7
websockets==11.0.3
